ProcessManager added error handling

+++ RemoteProcessManager/Managers/ProcessManager.cs
@@ -14,31 +14,49 @@ internal class ProcessManager : IProcessManager
 
     public void StartProcess(string processFullName, Action<string> onOutputData)
     {
-        if (_process?.HasExited is false)
+        if(File.Exists(processFullName) is false)
         {
-            _logger.LogWarning("Killing old process - ProcessId {ProcessId}", _process.Id);
-            _process.Kill();
-            _process.WaitForExit();
-            _process.Dispose();
+            _logger.LogWarning("Process file {processFullName}, not found.", processFullName);
+            onOutputData.Invoke($"Process file {processFullName}, not found.");
+            return;
         }
 
-        _logger.LogInformation("Starting process - {ProcessFullName}", processFullName);
-        _process = new Process();
-        _process.StartInfo.FileName = processFullName;
-        _process.StartInfo.UseShellExecute = false;
-        _process.StartInfo.RedirectStandardOutput = true;
-        _process.OutputDataReceived += (_, e) =>
+        try
         {
-            if (string.IsNullOrEmpty(e.Data)) return;
-            onOutputData.Invoke(e.Data);
-        };
+            if (_process?.HasExited is false)
+            {
+                _logger.LogWarning("Killing old process - ProcessId {ProcessId}", _process.Id);
+                onOutputData.Invoke($"Killing old process - ProcessId {_process.Id}");
+                _process.Kill();
+                _process.WaitForExit();
+                _process.Dispose();
+            }
 
-        _process.Start();
-        _logger.LogInformation("Process started - ProcessId {ProcessId}", _process.Id);
+            _logger.LogInformation("Starting process - {ProcessFullName}", processFullName);
+            onOutputData.Invoke($"Starting process - {processFullName}");
+            _process = new Process();
+            _process.StartInfo.FileName = processFullName;
+            _process.StartInfo.UseShellExecute = false;
+            _process.StartInfo.RedirectStandardOutput = true;
+            _process.OutputDataReceived += (_, e) =>
+            {
+                if (string.IsNullOrEmpty(e.Data)) return;
+                onOutputData.Invoke(e.Data);
+            };
 
-        // Asynchronously read the standard output of the spawned process.
-        // This raises OutputDataReceived events for each line of output.
-        _process.BeginOutputReadLine();
-        _process.WaitForExit();
+            _process.Start();
+            _logger.LogInformation("Process started - ProcessId {ProcessId}", _process.Id);
+            onOutputData.Invoke($"Process started - ProcessId {_process.Id}");
+
+            // Asynchronously read the standard output of the spawned process.
+            // This raises OutputDataReceived events for each line of output.
+            _process.BeginOutputReadLine();
+            _process.WaitForExit();
+        }
+        catch (Exception e)
+        {
+            _logger.LogError(e, "Start process Failed - {ProcessFullName}", processFullName);
+            onOutputData.Invoke($"Start process Failed - {processFullName}, Error: {e.Message}");
+        }
     }
 }
